Updated comments

+++ debt_marketplace/contracts/Loans.sol
@@ -3,10 +3,10 @@ pragma solidity ^0.8.0;
 
 contract Loans {
     struct Loan {
-        uint amount;
-        uint interestRate;
-        uint duration;
-        uint forSalePrice;
+        uint256 amount;
+        uint256 interestRate;
+        uint256 duration;
+        uint256 forSalePrice;
         bool isProposed;
         bool isActive;
         bool isForSale;
@@ -15,28 +15,26 @@ contract Loans {
     mapping(address => mapping(address => Loan)) public activeLoans;
 
     function proposeLoan(
-        uint _amount, 
-        uint _interesetRatePercent, 
-        uint _duration
-    ) 
-    external 
-    {
-        uint interestRate = 
-            (_amount * _interesetRatePercent * 10**18) / (100 * 10**18);
+        uint256 _amount,
+        uint256 _interesetRatePercent,
+        uint256 _duration
+    ) external {
+        uint256 interestRate = (_amount * _interesetRatePercent * 10**18) /
+            (100 * 10**18);
         require(
-            proposedLoans[msg.sender].isProposed == false, 
+            proposedLoans[msg.sender].isProposed == false,
             "Account already has proposed loan or has active loan"
         );
         proposedLoans[msg.sender] = Loan(
-            _amount, 
-            interestRate, 
-            _duration, 
+            _amount,
+            interestRate,
+            _duration,
             0,
             true, //isProposed
             false, //isActive
             false //isForSale
         );
-    }    
+    }
 
     function lend(address payable _borrower) public payable {
         //make sure loan exits and not active
@@ -50,34 +48,36 @@ contract Loans {
             proposedLoans[_borrower].interestRate,
             proposedLoans[_borrower].duration,
             0,
-            false, 
+            false,
             true,
             false
         );
+        //Deletes proposed loan mapping and updates it with who lent to the
+        // borrower using a nested mapping
         delete proposedLoans[_borrower];
-        uint amountToLend = activeLoans[msg.sender][_borrower].amount;
+        uint256 amountToLend = activeLoans[msg.sender][_borrower].amount;
+        //Transfers proposed loan amount from lender to borrower
         (bool success, ) = _borrower.call{value: amountToLend}("");
         require(success, "Transaction failed");
     }
 
+    //Borrower's call this function to pay back their loan
     function payback(address payable _lender) public payable {
-        uint debt = activeLoans[_lender][msg.sender].amount +
+        //Caluculates total borrower debt. Base loan + interest rate
+        uint256 debt = activeLoans[_lender][msg.sender].amount +
             activeLoans[_lender][msg.sender].interestRate;
         require(
-            activeLoans[_lender][msg.sender].isActive == true, 
+            activeLoans[_lender][msg.sender].isActive == true,
             "Nonexistant loan cannot be paid back"
         );
+        //Ensures the amount of ETH paid back matches how much is owed
         require(msg.value == debt, "Amount paid back has to be exact");
         (bool success, ) = _lender.call{value: debt}("");
         require(success, "Transaction failed");
     }
 
-    function listLoan(
-        address _borrower,
-        uint _salePrice
-    ) 
-        external 
-    {
+    //Lenders can call this function to sell off their loan to someone else
+    function listLoan(address _borrower, uint256 _salePrice) external {
         require(
             activeLoans[msg.sender][_borrower].isActive == true,
             "You do not have the rights to sell this loan"
@@ -86,14 +86,12 @@ contract Loans {
         activeLoans[msg.sender][_borrower].isForSale = true;
     }
 
-    function buyLoan(
-        address payable _lender, 
-        address _borrower
-    ) 
-        external payable
+    function buyLoan(address payable _lender, address _borrower)
+        external
+        payable
     {
         require(
-            activeLoans[_lender][_borrower].isForSale == true && 
+            activeLoans[_lender][_borrower].isForSale == true &&
                 msg.value == activeLoans[_lender][_borrower].forSalePrice,
             "Loan either does not exist or is not for sale"
         );
